fix(quality): zero out repo-wide lint/test debt and gate every surface (#201)
Fixes #201

Python (was: no linter wired up at all — 562 raw ruff findings):
- ruff.toml: default F + E4/E7/E9 rules; excludes vendored layer build
  output (lambda/layers/*/python — 460 of the 562 were bs4/six/etc.)
- 102 first-party errors fixed for real: 82 unused imports removed
  (full suite passed after — none were load-bearing), 7 pointless
  f-strings, 3 redefined-while-unused, 3 unused variables (two test
  'body' assignments became real response-body assertions; one dead
  'token' deleted), 7 E402s fixed structurally (processor's sys.path
  extension collapsed to the single-call form ruff exempts; stray
  mid-file 'import logging' hoisted in test_play_client)
- processor/handler.py: MessageValidationError dropped from the
  availability probe import (unused)
- requirements-dev.txt: black/isort/mypy were declared but nothing ever
  ran them (repo-wide grep: zero references) — replaced with ruff, the
  linter the gate now actually enforces

Tests:
- pytest.ini: testpaths = lambda plugins — ~205 plugin tests were
  excluded from every run (--ignore=plugins + norecursedirs); coverage
  now spans both trees. Root test:backend drops the hardcoded lambda/
  arg. Full run: 1096 passed, 0 skipped.
- lambda/shared/prompts.py: local-dev prompts resolution pointed at a
  directory that does not exist (parent.parent.parent/'prompts');
  fixed to lambda/api/prompts — and test_real_resolution no longer
  hides that regression behind pytest.skip (hard assertion now).

Frontend (3 react-refresh warnings, invisible because warnings never
failed the gate):
- describeRouteError extracted from RouteErrorBoundary/index.tsx to its
  own module; normalizedBaseUrl + scraperDomainLabel extracted from
  ScraperCard.tsx to scraperUrl.ts; tests import from the new homes
- frontend lint script: eslint . --max-warnings 0 — warnings fail from
  now on

Gate: root lint = frontend + stream + python (lint:python script);
test:backend includes plugins. Mutation-verified: an unused import
fails lint:python, a react-refresh warning fails lint:frontend, a
broken plugin test fails test:backend. Full npm run check: exit 0.
