+++ package.json
@@ -8,16 +8,17 @@
     "build:layers": "cd voc-datalake && chmod +x scripts/build-layers.sh && ./scripts/build-layers.sh",
     "dev": "npm run dev --prefix voc-datalake/frontend",
     "build": "npm run build --prefix voc-datalake/frontend",
-    "lint": "npm run lint:frontend && npm run lint:stream",
+    "lint": "npm run lint:frontend && npm run lint:stream && npm run lint:python",
     "lint:frontend": "npm run lint --prefix voc-datalake/frontend",
     "lint:stream": "npm run lint --prefix voc-datalake/lambda/stream",
+    "lint:python": "cd voc-datalake && .venv/bin/ruff check lambda plugins",
     "typecheck": "npm run typecheck --prefix voc-datalake/frontend",
     "typecheck:tests": "npm run typecheck:tests --prefix voc-datalake/frontend",
     "typecheck:cdk": "cd voc-datalake && npx tsc --noEmit",
     "typecheck:stream": "cd voc-datalake/lambda/stream && npx tsc --noEmit",
     "typecheck:all": "npm run typecheck && npm run typecheck:cdk && npm run typecheck:stream",
     "test": "npm run test --prefix voc-datalake/frontend -- --run",
-    "test:backend": "cd voc-datalake && .venv/bin/python -m pytest lambda/ -q --no-header --tb=short",
+    "test:backend": "cd voc-datalake && .venv/bin/python -m pytest -q --no-header --tb=short",
     "test:cdk": "npm run test --prefix voc-datalake",
     "test:stream": "npm run test --prefix voc-datalake/lambda/stream",
     "test:watch": "npm run test:watch --prefix voc-datalake/frontend",
