+++ voc-datalake/frontend/package.json
@@ -7,7 +7,7 @@
     "dev": "vite",
     "prebuild": "cd .. && npm run generate:manifests && npm run generate:menu",
     "build": "tsc -b && vite build",
-    "lint": "eslint .",
+    "lint": "eslint . --max-warnings 0",
     "preview": "vite preview",
     "mock": "node mock-server.js",
     "test": "vitest",
