+++ voc-datalake/frontend/src/components/RouteErrorBoundary/RouteErrorBoundary.test.tsx
@@ -8,7 +8,8 @@ import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest'
 import { render, screen } from '@testing-library/react'
 import userEvent from '@testing-library/user-event'
 import { createMemoryRouter, Outlet, RouterProvider } from 'react-router-dom'
-import RouteErrorBoundary, { describeRouteError } from './index'
+import RouteErrorBoundary from './index'
+import { describeRouteError } from './describeRouteError'
 
 function TestLayout() {
   return (
