+++ voc-datalake/frontend/src/components/RouteErrorBoundary/index.tsx
@@ -10,20 +10,8 @@
 import { useEffect } from 'react'
 import { AlertTriangle, Home, RotateCcw } from 'lucide-react'
 import { useTranslation } from 'react-i18next'
-import { isRouteErrorResponse, Link, useRouteError } from 'react-router-dom'
-
-/** Human-readable summary of whatever useRouteError() delivered (unknown). */
-export function describeRouteError(error: unknown): string {
-  if (isRouteErrorResponse(error)) {
-    const base = `${error.status} ${error.statusText}`.trim()
-    // Loader-thrown Responses often carry the actionable message in data.
-    return typeof error.data === 'string' && error.data ? `${base} — ${error.data}` : base
-  }
-  if (error instanceof Error) {
-    return error.message
-  }
-  return String(error)
-}
+import { Link, useRouteError } from 'react-router-dom'
+import { describeRouteError } from './describeRouteError'
 
 export default function RouteErrorBoundary() {
   const error = useRouteError()
