+++ voc-datalake/lambda/api/settings_handler.py
@@ -15,7 +15,7 @@ from botocore.exceptions import ClientError
 # Add shared module to path
 sys.path.insert(0, os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 
-from shared.logging import logger, tracer, metrics
+from shared.logging import logger, tracer
 from shared.aws import get_dynamodb_resource
 from shared.api import create_api_resolver, api_handler, require_admin
 from shared.exceptions import ConfigurationError, ValidationError, ServiceError
