+++ voc-datalake/lambda/api/test/test_feedback_form_handler.py
@@ -2,9 +2,7 @@
 Tests for feedback_form_handler.py - /feedback-forms/* endpoints.
 """
 import json
-import pytest
-from unittest.mock import patch, MagicMock
-from datetime import datetime, timezone
+from unittest.mock import patch
 
 
 class TestListForms:
