+++ voc-datalake/lambda/api/test/test_projects_handler.py
@@ -2,9 +2,7 @@
 Tests for projects_handler.py - /projects/* endpoints.
 """
 import json
-import pytest
-from unittest.mock import patch, MagicMock
-from datetime import datetime, timezone
+from unittest.mock import patch
 
 
 class TestValidatePersonaCount:
@@ -144,6 +142,7 @@ class TestGetProjectEndpoint:
         body = json.loads(response['body'])
         
         mock_get_project.assert_called_once_with('proj-123')
+        assert body['project']['project_id'] == 'proj-123'
 
 
 class TestUpdateProjectEndpoint:
