+++ voc-datalake/lambda/jobs/document_generator/handler.py
@@ -6,7 +6,6 @@ Uses the same prompt templates and chain patterns as the synchronous projects.py
 """
 
 import os
-import re
 import sys
 from datetime import datetime, timezone
