+++ voc-datalake/plugins/_shared/test/test_circuit_breaker.py
@@ -1,8 +1,5 @@
 """Tests for circuit_breaker.py"""
-import os
-import pytest
 from unittest.mock import patch, MagicMock
-from datetime import datetime, timezone, timedelta
 
 
 class TestCircuitBreakerInit:
