+++ voc-datalake/plugins/test_synthetic_reviews.py
@@ -8,7 +8,6 @@ import json
 from contextlib import contextmanager
 from unittest.mock import patch, MagicMock
 
-import pytest
 
 
 @contextmanager
