+++ voc-datalake/pytest.ini
@@ -1,9 +1,9 @@
 [pytest]
-testpaths = lambda
+testpaths = lambda plugins
 python_files = test_*.py *_test.py
 python_classes = Test*
 python_functions = test_*
-addopts = -v --tb=short --ignore=cdk.out --ignore=plugins --ignore=dist --ignore=frontend --ignore=lambda/layers --cov=lambda --cov-report=term-missing --cov-report=html:coverage_html --cov-fail-under=0
+addopts = -v --tb=short --ignore=cdk.out --ignore=dist --ignore=frontend --ignore=lambda/layers --ignore=plugins/_template --cov=lambda --cov=plugins --cov-report=term-missing --cov-report=html:coverage_html --cov-fail-under=0
 filterwarnings =
     ignore::DeprecationWarning
     ignore::PendingDeprecationWarning
@@ -11,4 +11,4 @@ filterwarnings =
 markers =
     slow: marks tests as slow (deselect with '-m "not slow"')
     integration: marks tests as integration tests
-norecursedirs = cdk.out .venv node_modules dist frontend plugins .git layers
+norecursedirs = cdk.out .venv node_modules dist frontend .git layers _template
