+++ voc-datalake/requirements-dev.txt
@@ -14,7 +14,6 @@ boto3-stubs[dynamodb,s3,sqs,secretsmanager,bedrock-runtime]>=1.35.0
 # Lambda Powertools (same as production)
 aws-lambda-powertools>=3.0.0
 
-# Code quality
-black>=24.0.0
-isort>=5.13.0
-mypy>=1.11.0
+# Code quality — ruff is the only Python linter; the root gate runs
+# `ruff check lambda plugins` (config in ruff.toml, issue #201)
+ruff>=0.15.0
