+++ voc-datalake/ruff.toml
@@ -1,0 +1,15 @@
+# Python lint configuration (issue #201).
+#
+# Rules are ruff's defaults (pyflakes F + pycodestyle errors E4/E7/E9) — the
+# baseline every first-party file passes. The root gate runs
+# `ruff check lambda plugins` from this directory; keep it at zero.
+
+extend-exclude = [
+    # Vendored third-party code: local layer build output from
+    # scripts/build-layers.sh (bs4, six, ...). Never lint, never fix.
+    "lambda/layers",
+    "cdk.out",
+    "node_modules",
+    "frontend",
+    "dist",
+]
